Tamarin has a supported Win64/x86-64 release, but we need Unix/x86-64. This is a snapshot of the work on that port, mainly implementing the System V Application Binary Interface, and fixing a few Adobe bugs:  - The current code correctly interpets all Tamarin acceptance tests except one (*)  - No JIT support yet (lots of work in CodegenMIR)  - Broken code emitting for interfaces (Amd64Assembler::emitImtThunk)  - No debugger support (mostly needs work in Amd64Assembler)
(*) The failing acceptance test executes a native method that takes more than six arguments. It may seem odd that this should matter; the reason it does is that the x86_64 ABI passes 6 integer type arguments by register before resorting to stack, and the stack based passing is still being ornery and is disabled.