Well, that was fun.  I always wanted to get into thane development, Zell has been swamped, and the annoying server-side usercode buffering was getting in the way of UD progress, so I decided to dive in.
In this new version of thane, they abstracted a bunch of stuff into a Platform interface.  There is
a Win implementation and a PosixPartial implementation (which then has further subclasses I didn't
need to worry about).  I'm not entirely sure why they moved stdout logging to Platform, because
the implementation is identical in Win and PosixPartial.  I am relatively sure though, that that
is where the issue with not flushing stdout came in.

Solution: Add a flush() method to their OutputStream interface and to their PrintWriter interface.
Use that method in SystemClass::trace(), which is where all user-code traces end up.  I never was
able to determine why stdout is getting buffered.  Is that common?  I don't remember, but it does
seem like a good idea because in my testing I saw that it's getting spammed, usually with single
characters, when traces go through the path from SystemClass to stdout.  So, whether the buffering
is new or not, I left whatever is causing it alone and added flushing in the one place we really
want it.

This is only tested to build and work on Linux-32.  Zell, I assume you'll want to drive it around
on Windows and Mac.  Let me know if you want me to commit my Linux-32 build.